requirements.txt
streamlit
yt-dlp
moviepy
opencv-python
scikit-image
imagehash
Pillow